+++ firmware/controllers/trigger/decoders/trigger_structure.cpp
@@ -550,11 +550,16 @@ void TriggerWaveform::initializeTriggerWaveform(operation_mode_e triggerOperatio
 		setMercedesTwoSegment(this);
 		break;
 
-	case trigger_type_e::TT_ONE:
-		/** @note TT_ONE setup events as 180 and 360 degrees. It uses SyncEdge::Rise
-		 *  for additionaly phase align on falling edge and will not work with non-
-		 *  symmetrical blind type where open and closed sections are not qual */
+	case trigger_type_e::TT_HALF_MOON:
+		/** @note TT_HALF_MOON setup events as 180 and 360 degrees. It uses SyncEdge::Rise
+		 *  for additional phase align on falling edge and will not work with non-
+		 *  symmetrical blind type where open and closed sections are not equal */
 		initializeSkippedToothTrigger(this, 1, 0, triggerOperationMode, SyncEdge::Rise);
+		break;
+
+	case trigger_type_e::TT_NARROW_SINGLE_TOOTH:
+		/** different from TT_HALF_MOON */
+		initializeSkippedToothTrigger(this, 1, 0, triggerOperationMode, SyncEdge::RiseOnly);
 		break;
 
 	case trigger_type_e::TT_MAZDA_SOHC_4:
@@ -646,7 +651,6 @@ void TriggerWaveform::initializeTriggerWaveform(operation_mode_e triggerOperatio
 		 * just to use another enum for each trigger type. */
 		configure60degSingleTooth(this);
 		break;
-	case trigger_type_e::TT_UNUSED_11:
 	case trigger_type_e::TT_UNUSED_75:
 	case trigger_type_e::TT_UNUSED_78:
 	case trigger_type_e::TT_MITSU_4G63_CAM: